+++ src/data/protocols/http/http-post-client.ts
@@ -1,4 +1,4 @@
-import { HttpResponse } from './http-response'
+import { HttpResponse } from '.'
 
 export type HttpPostParams<T> = {
   url: string
