+++ src/data/usecases/authentication/remote-authentication.spec.ts
@@ -1,11 +1,10 @@
-import { HttpPostClientSpy } from 'data/test/mock-http-client'
-import { InvalidCredentialsError } from 'domain/errors/invalid-credentials-error'
-import { mockAccountModel, mockAuthentication } from 'domain/test/mock-account'
 import { RemoteAuthentication } from './remote-authentication'
-import { HttpStatusCode } from 'data/protocols/http/http-response'
-import { AuthenticationParams } from 'domain/usecases/authentication'
-import { AccountModel } from 'domain/models/account-model'
-import { UnexpectedError } from 'domain/errors/unexpected-error'
+import { HttpPostClientSpy } from 'data/test'
+import { UnexpectedError, InvalidCredentialsError } from 'domain/errors'
+import { mockAccountModel, mockAuthentication } from 'domain/test'
+import { HttpStatusCode } from 'data/protocols/http'
+import { AuthenticationParams } from 'domain/usecases'
+import { AccountModel } from 'domain/models'
 import faker from 'faker'
 
 type SutTypes = {
