+++ src/data/usecases/authentication/remote-authentication.ts
@@ -1,9 +1,7 @@
-import { HttpPostClient } from 'data/protocols/http/http-post-client'
-import { HttpStatusCode } from 'data/protocols/http/http-response'
-import { Authentication, AuthenticationParams } from 'domain/usecases/authentication'
-import { AccountModel } from 'domain/models/account-model'
-import { InvalidCredentialsError } from 'domain/errors/invalid-credentials-error'
-import { UnexpectedError } from 'domain/errors/unexpected-error'
+import { HttpPostClient, HttpStatusCode } from 'data/protocols/http'
+import { Authentication, AuthenticationParams } from 'domain/usecases'
+import { AccountModel } from 'domain/models'
+import { UnexpectedError, InvalidCredentialsError } from 'domain/errors'
 
 export class RemoteAuthentication implements Authentication {
   constructor (
